Remove package helpful
In a recent commit a dependency towards shut-up was added, which
imo (having seen the pull request) wasn't needed, as this was a hacky
solution to an underlying problem that could be addressed differently.
Since I don't use helpful all that much though, I don't care enough to
keep it, so I removed it, as I hate extraneous dependencies.