:wrench: Shorten link of badge with logo
In the original link, the logo was read from a base64 string, and it
was thus rather long. Turns out I had underestimated its length, and
it was long enough to occupy half of the page when editing in source
view, something rather annoying. I thus shortened the link via
frama.link. Of course this means I'll have to reshorten it when the
version of the editor changes, but that won't happen too often.