:warning: :wrench: Set init.el to be tangled from README.org
I will be keeping init.el in version control though, because that will
make it easier to bootstrap the config on new systems. I also set the
init.el to be tangled from the org file automatically upon saving, but
this takes a few seconds every time, so not sure if I'll be keeping that.