:wrench: Change magit installation trigger
It now detects whether git is installed instead of if the platform is
linux. This will not only allow magit to be installed on non-linux
systems with git installed, but also will prevent its installation in
linux systems that might not have git installed. I've read that
supposedly magit is slow on windows. If that is the case, I'll just
remove git for windows from the universal path, since I otherwise use
it only through cmder, which detects it by itself (maybe it uses its
own path variables).