:wrench: Tweak some ryo modal keys
I finally figured out exactly what ':norepeat' does (hint: it's
similar to the vim behavior). This is going to be useful for repeating
marking lines, as I want to keep the line-move-visual behavior of j
and k for the majority of cases and don't want it only when marking
lines, which can be done the intended way by repeating V, and this
becomes easier by using ryo-modal-repeat, here bound to the dot symbol
(that is, no need to press Shift+v when repeating marking lines now -
now I can just press dot).